save in h5

+++ utils/general.py
@@ -34,7 +34,7 @@ def train(model, trainLoader, optimizer, loss_function, device, writer, step):
     model.train()
     for batch, (druginputBatch, targetinputBatch, labelinputBatch) in enumerate(tqdm(trainLoader, leave=False, desc="Train", ncols=75)):
         druginputBatch = (druginputBatch[0].float().to(device), druginputBatch[1].float().to(device), druginputBatch[2].bool().to(device))
-        targetinputBatch = (targetinputBatch[0].int().to(device), targetinputBatch[1].bool().to(device))
+        targetinputBatch = (targetinputBatch[0].to(device), targetinputBatch[1].bool().to(device))
         labelinputBatch = labelinputBatch.long().to(device)
 
         optimizer.zero_grad()
@@ -63,7 +63,7 @@ def evaluate(model, evalLoader, loss_function, device):
 
     for batch, (druginputBatch, targetinputBatch, labelinputBatch) in enumerate(tqdm(evalLoader, leave=False, desc="Eval", ncols=75)):
         druginputBatch = (druginputBatch[0].float().to(device), druginputBatch[1].float().to(device), druginputBatch[2].bool().to(device))
-        targetinputBatch = (targetinputBatch[0].int().to(device), targetinputBatch[1].bool().to(device))
+        targetinputBatch = (targetinputBatch[0].to(device), targetinputBatch[1].bool().to(device))
         labelinputBatch = labelinputBatch.long().to(device)
 
         model.eval()
